+++ kpops/component_handlers/kafka_connect/connect_wrapper.py
@@ -19,9 +19,7 @@ log = logging.getLogger("KafkaConnectAPI")
 
 
 class ConnectWrapper:
-    """
-    Wraps Kafka Connect APIs
-    """
+    """Wraps Kafka Connect APIs."""
 
     def __init__(self, host: str | None):
         if not host:
@@ -40,11 +38,11 @@ class ConnectWrapper:
     async def create_connector(
         self, connector_config: KafkaConnectorConfig
     ) -> KafkaConnectResponse:
-        """
-        Creates a new connector
+        """Create a new connector.
+
         API Reference: https://docs.confluent.io/platform/current/connect/references/restapi.html#post--connectors
         :param connector_config: The config of the connector
-        :return: The current connector info if successful
+        :return: The current connector info if successful.
         """
 
         config_json = connector_config.dict()
@@ -67,11 +65,13 @@ class ConnectWrapper:
         raise KafkaConnectError(response)
 
     async def get_connector(self, connector_name: str) -> KafkaConnectResponse:
-        """
-        Get information about the connector.
-        API Reference: https://docs.confluent.io/platform/current/connect/references/restapi.html#get--connectors-(string-name)
+        """Get information about the connector.
+
+        API Reference:
+        https://docs.confluent.io/platform/current/connect/references/restapi.html#get--connectors-(string-name)
+
         :param connector_name: Nameof the crated connector
-        :return: Information about the connector
+        :return: Information about the connector.
         """
         response = await self._client.get(
             url=f"/connectors/{connector_name}", headers=HEADERS
@@ -82,7 +82,7 @@ class ConnectWrapper:
             return KafkaConnectResponse(**response.json())
         elif response.status_code == httpx.codes.NOT_FOUND:
             log.info(f"The named connector {connector_name} does not exists.")
-            raise ConnectorNotFoundException()
+            raise ConnectorNotFoundException
         elif response.status_code == httpx.codes.CONFLICT:
             log.warning(
                 "Rebalancing in progress while getting a connector... Retrying..."
@@ -94,8 +94,11 @@ class ConnectWrapper:
     async def update_connector_config(
         self, connector_config: KafkaConnectorConfig
     ) -> KafkaConnectResponse:
-        """
-        Create a new connector using the given configuration, or update the configuration for an existing connector.
+        """Create or update a connector.
+
+        Create a new connector using the given configuration,or update the
+        configuration for an existing connector.
+
         :param connector_config: Configuration parameters for the connector.
         :return: Information about the connector after the change has been made.
         """
@@ -128,10 +131,11 @@ class ConnectWrapper:
     async def validate_connector_config(
         self, connector_config: KafkaConnectorConfig
     ) -> list[str]:
-        """
-        Validate connector config using the given configuration
+        """Validate connector config using the given configuration.
+
         :param connector_config: Configuration parameters for the connector.
-        :return:
+        :raises KafkaConnectError: Kafka Konnect error
+        :return: List of all found errors
         """
 
         response = await self._client.put(
@@ -145,7 +149,7 @@ class ConnectWrapper:
                 **response.json()
             )
 
-            errors = []
+            errors: list[str] = []
             if kafka_connect_error_response.error_count > 0:
                 for config in kafka_connect_error_response.configs:
                     if len(config.value.errors) > 0:
@@ -157,9 +161,12 @@ class ConnectWrapper:
         raise KafkaConnectError(response)
 
     async def delete_connector(self, connector_name: str) -> None:
-        """
-        Deletes a connector, halting all tasks and deleting its configuration.
-        API Reference:https://docs.confluent.io/platform/current/connect/references/restapi.html#delete--connectors-(string-name)-
+        """Delete a connector, halting all tasks and deleting its configuration.
+
+        API Reference:
+            https://docs.confluent.io/platform/current/connect/references/restapi.html#delete--connectors-(string-name)-.
+        :param connector_name: Configuration parameters for the connector.
+        :raises ConnectorNotFoundException: Connector not found
         """
         response = await self._client.delete(
             url=f"/connectors/{connector_name}", headers=HEADERS
@@ -169,7 +176,7 @@ class ConnectWrapper:
             return
         elif response.status_code == httpx.codes.NOT_FOUND:
             log.info(f"The named connector {connector_name} does not exists.")
-            raise ConnectorNotFoundException()
+            raise ConnectorNotFoundException
         elif response.status_code == httpx.codes.CONFLICT:
             log.warning(
                 "Rebalancing in progress while deleting a connector... Retrying..."