+++ kpops/component_handlers/topic/proxy_wrapper.py
@@ -21,15 +21,12 @@ HEADERS = {"Content-Type": "application/json"}
 
 
 class ProxyWrapper:
-    """
-    Wraps Kafka REST Proxy APIs
-    """
+    """Wraps Kafka REST Proxy APIs."""
 
     def __init__(self, pipeline_config: PipelineConfig) -> None:
         if not pipeline_config.kafka_rest_host:
-            raise ValueError(
-                "The Kafka REST Proxy host is not set. Please set the host in the config.yaml using the kafka_rest_host property or set the environemt variable KPOPS_REST_PROXY_HOST."
-            )
+            msg = "The Kafka REST Proxy host is not set. Please set the host in the config.yaml using the kafka_rest_host property or set the environemt variable KPOPS_REST_PROXY_HOST."
+            raise ValueError(msg)
         self._client = httpx.AsyncClient(
             base_url=f"{pipeline_config.kafka_rest_host}/v3/clusters"
         )
@@ -40,13 +37,15 @@ class ProxyWrapper:
 
     @cached_property
     def cluster_id(self) -> str:
-        """
-        Gets the Kafka cluster ID by sending a requests to Kafka REST proxy.
+        """Get the Kafka cluster ID by sending a request to Kafka REST proxy.
+
         More information about the cluster ID can be found here:
-        https://docs.confluent.io/platform/current/kafka-rest/api.html#cluster-v3
+        https://docs.confluent.io/platform/current/kafka-rest/api.html#cluster-v3.
 
         Currently both Kafka and Kafka REST Proxy are only aware of the Kafka cluster pointed at by the
         bootstrap.servers configuration. Therefore, only one Kafka cluster will be returned.
+
+        :raises KafkaRestProxyError: Kafka REST proxy error
         :return: The Kafka cluster ID.
         """
         response = self._sync_client.get("")
@@ -62,10 +61,13 @@ class ProxyWrapper:
         return self._host
 
     async def create_topic(self, topic_spec: TopicSpec) -> None:
-        """
-        Creates a topic.
-        API Reference: https://docs.confluent.io/platform/current/kafka-rest/api.html#post--clusters-cluster_id-topics
+        """Create a topic.
+
+        API Reference:
+        https://docs.confluent.io/platform/current/kafka-rest/api.html#post--clusters-cluster_id-topics
+
         :param topic_spec: The topic specification.
+        :raises KafkaRestProxyError: Kafka REST proxy error
         """
         response = await self._client.post(
             url=f"/{self.cluster_id}/topics",
@@ -81,10 +83,13 @@ class ProxyWrapper:
         raise KafkaRestProxyError(response)
 
     async def delete_topic(self, topic_name: str) -> None:
-        """
-        Deletes a topic
-        API Reference: https://docs.confluent.io/platform/current/kafka-rest/api.html#delete--clusters-cluster_id-topics-topic_name
-        :param topic_name: Name of the topic
+        """Delete a topic.
+
+        API Reference:
+        https://docs.confluent.io/platform/current/kafka-rest/api.html#delete--clusters-cluster_id-topics-topic_name
+
+        :param topic_name: Name of the topic.
+        :raises KafkaRestProxyError: Kafka REST proxy error
         """
         response = await self._client.delete(
             url=f"/{self.cluster_id}/topics/{topic_name}",
@@ -98,11 +103,15 @@ class ProxyWrapper:
         raise KafkaRestProxyError(response)
 
     async def get_topic(self, topic_name: str) -> TopicResponse:
-        """
-        Returns the topic with the given topic_name.
-        API Reference: https://docs.confluent.io/platform/current/kafka-rest/api.html#get--clusters-cluster_id-topics-topic_name
+        """Return the topic with the given topic_name.
+
+        API Reference:
+        https://docs.confluent.io/platform/current/kafka-rest/api.html#get--clusters-cluster_id-topics-topic_name
+
         :param topic_name: The topic name.
-        :return: Response of the get topic API
+        :raises TopicNotFoundException: Topic not found
+        :raises KafkaRestProxyError: Kafka REST proxy error
+        :return: Response of the get topic API.
         """
 
         response = await self._client.get(
@@ -121,15 +130,19 @@ class ProxyWrapper:
         ):
             log.debug(f"Topic {topic_name} not found.")
             log.debug(response.json())
-            raise TopicNotFoundException()
+            raise TopicNotFoundException
 
         raise KafkaRestProxyError(response)
 
     async def get_topic_config(self, topic_name: str) -> TopicConfigResponse:
-        """
-        Return the config with the given topic_name.
-        API Reference: https://docs.confluent.io/platform/current/kafka-rest/api.html#acl-v3
+        """Return the config with the given topic_name.
+
+        API Reference:
+        https://docs.confluent.io/platform/current/kafka-rest/api.html#acl-v3
+
         :param topic_name: The topic name.
+        :raises TopicNotFoundException: Topic not found
+        :raises KafkaRestProxyError: Kafka REST proxy error
         :return: The topic configuration.
         """
 
@@ -149,18 +162,21 @@ class ProxyWrapper:
         ):
             log.debug(f"Configs for {topic_name} not found.")
             log.debug(response.json())
-            raise TopicNotFoundException()
+            raise TopicNotFoundException
 
         raise KafkaRestProxyError(response)
 
     async def batch_alter_topic_config(
         self, topic_name: str, json_body: list[dict]
     ) -> None:
-        """
-        Reset config of given config_name param to the default value on the kafka server.
-        API Reference: https://docs.confluent.io/platform/current/kafka-rest/api.html#post--clusters-cluster_id-topics-topic_name-configs-alter
+        """Reset config of given config_name param to the default value on the kafka server.
+
+        API Reference:
+        https://docs.confluent.io/platform/current/kafka-rest/api.html#post--clusters-cluster_id-topics-topic_name-configs-alter
+
         :param topic_name: The topic name.
         :param config_name: The configuration parameter name.
+        :raises KafkaRestProxyError: Kafka REST proxy error
         """
         response = await self._client.post(
             url=f"/{self.cluster_id}/topics/{topic_name}/configs:alter",
@@ -175,9 +191,12 @@ class ProxyWrapper:
         raise KafkaRestProxyError(response)
 
     async def get_broker_config(self) -> BrokerConfigResponse:
-        """
-        Return the list of configuration parameters for all the brokers in the given Kafka cluster.
-        API Reference: https://docs.confluent.io/platform/current/kafka-rest/api.html#get--clusters-cluster_id-brokers---configs
+        """Return the list of configuration parameters for all the brokers in the given Kafka cluster.
+
+        API Reference:
+        https://docs.confluent.io/platform/current/kafka-rest/api.html#get--clusters-cluster_id-brokers---configs
+
+        :raises KafkaRestProxyError: Kafka REST proxy error
         :return: The broker configuration.
         """
         response = await self._client.get(