+++ kpops/utils/environment.py
@@ -1,36 +1,42 @@
 import os
-import platform
 from collections import UserDict
-from typing import Callable
+from collections.abc import ItemsView, KeysView, MutableMapping, ValuesView
 
 
-class Environment(UserDict):
-    def __init__(self, mapping=None, /, **kwargs) -> None:
-        transformation = Environment.__get_transformation()
-        if mapping is not None:
-            mapping = {transformation(key): value for key, value in mapping.items()}
-        else:
+class Environment(UserDict[str, str]):
+    """Internal environment wrapping OS environment."""
+
+    def __init__(
+        self, mapping: MutableMapping[str, str] | None = None, /, **kwargs: str
+    ) -> None:
+        self._global = os.environ
+        if mapping is None:
             mapping = {}
         if kwargs:
-            mapping.update(
-                {transformation(key): value for key, value in kwargs.items()}
-            )
+            mapping.update(**kwargs)
         super().__init__(mapping)
 
-    @staticmethod
-    def __key_camel_case_transform(key: str) -> str:
-        return key.lower()
+    def __getitem__(self, key: str) -> str:
+        try:
+            return self.data[key]
+        except KeyError:
+            return self._global[key]
 
-    @staticmethod
-    def __key_identity_transform(key: str) -> str:
-        return key
+    def __contains__(self, key: object) -> bool:
+        return super().__contains__(key) or self._global.__contains__(key)
 
-    @staticmethod
-    def __get_transformation() -> Callable[[str], str]:
-        if platform.system() == "Windows":
-            return Environment.__key_camel_case_transform
-        else:
-            return Environment.__key_identity_transform
+    @property
+    def _dict(self) -> dict[str, str]:
+        return {**self._global, **self.data}
+
+    def keys(self) -> KeysView[str]:
+        return KeysView(self._dict)
+
+    def values(self) -> ValuesView[str]:
+        return ValuesView(self._dict)
+
+    def items(self) -> ItemsView[str, str]:
+        return ItemsView(self._dict)
 
 
-ENV = Environment(os.environ)
+ENV = Environment()