+++ tests/component_handlers/kafka_connect/test_connect_handler.py
@@ -22,19 +22,19 @@ TOPIC_NAME = "test-topic"
 
 
 class TestConnectorHandler:
-    @pytest.fixture
+    @pytest.fixture()
     def log_info_mock(self, mocker: MockerFixture) -> MagicMock:
         return mocker.patch(
             "kpops.component_handlers.kafka_connect.kafka_connect_handler.log.info"
         )
 
-    @pytest.fixture
+    @pytest.fixture()
     def log_warning_mock(self, mocker: MockerFixture) -> MagicMock:
         return mocker.patch(
             "kpops.component_handlers.kafka_connect.kafka_connect_handler.log.warning"
         )
 
-    @pytest.fixture
+    @pytest.fixture()
     def log_error_mock(self, mocker: MockerFixture) -> MagicMock:
         return mocker.patch(
             "kpops.component_handlers.kafka_connect.kafka_connect_handler.log.error"
@@ -44,7 +44,7 @@ class TestConnectorHandler:
     def connector_wrapper(self) -> AsyncMock:
         return AsyncMock()
 
-    @pytest.fixture
+    @pytest.fixture()
     def renderer_diff_mock(self, mocker: MockerFixture) -> MagicMock:
         return mocker.patch(
             "kpops.component_handlers.kafka_connect.kafka_connect_handler.render_diff"
@@ -57,7 +57,7 @@ class TestConnectorHandler:
             timeout=0,
         )
 
-    @pytest.fixture
+    @pytest.fixture()
     def connector_config(self) -> KafkaConnectorConfig:
         return KafkaConnectorConfig(
             **{