+++ tests/component_handlers/kafka_connect/test_connect_wrapper.py
@@ -27,7 +27,7 @@ DEFAULTS_PATH = Path(__file__).parent / "resources"
 
 class TestConnectorApiWrapper:
     @pytest_asyncio.fixture(autouse=True)
-    async def setup(self):
+    def _setup(self):
         config = PipelineConfig(
             defaults_path=DEFAULTS_PATH,
             environment="development",
@@ -35,7 +35,7 @@ class TestConnectorApiWrapper:
         )
         self.connect_wrapper = ConnectWrapper(host=config.kafka_connect_host)
 
-    @pytest.fixture
+    @pytest.fixture()
     def connector_config(self) -> KafkaConnectorConfig:
         return KafkaConnectorConfig(
             **{
@@ -535,9 +535,9 @@ class TestConnectorApiWrapper:
 
     @pytest.mark.asyncio
     async def test_should_parse_validate_connector_config(self, httpx_mock: HTTPXMock):
-        with open(
+        with Path(
             DEFAULTS_PATH / "connect_validation_response.json",
-        ) as f:
+        ).open() as f:
             actual_response = json.load(f)
 
         httpx_mock.add_response(