+++ tests/components/test_kafka_source_connector.py
@@ -24,14 +24,15 @@ from kpops.components.base_components.models.to_section import (
 )
 from kpops.utils.environment import ENV
 from tests.components.test_kafka_connector import (
-    CONNECTOR_CLEAN_NAME,
+    CONNECTOR_CLEAN_FULL_NAME,
+    CONNECTOR_FULL_NAME,
     CONNECTOR_NAME,
     TestKafkaConnector,
 )
 
 
 class TestKafkaSourceConnector(TestKafkaConnector):
-    @pytest.fixture
+    @pytest.fixture()
     def connector(
         self,
         config: PipelineConfig,
@@ -114,7 +115,9 @@ class TestKafkaSourceConnector(TestKafkaConnector):
 
         await connector.destroy(dry_run=True)
 
-        mock_destroy_connector.assert_called_once_with(CONNECTOR_NAME, dry_run=True)
+        mock_destroy_connector.assert_called_once_with(
+            CONNECTOR_FULL_NAME, dry_run=True
+        )
 
     @pytest.mark.asyncio
     async def test_reset_when_dry_run_is_true(
@@ -158,11 +161,11 @@ class TestKafkaSourceConnector(TestKafkaConnector):
             ),
             mocker.call.helm.uninstall(
                 namespace="test-namespace",
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 dry_run=False,
             ),
             mocker.call.helm.upgrade_install(
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 namespace="test-namespace",
                 chart="bakdata-kafka-connect-resetter/kafka-connect-resetter",
                 dry_run=False,
@@ -175,15 +178,15 @@ class TestKafkaSourceConnector(TestKafkaConnector):
                     "connectorType": "source",
                     "config": {
                         "brokers": "broker:9092",
-                        "connector": CONNECTOR_NAME,
+                        "connector": CONNECTOR_FULL_NAME,
                         "offsetTopic": "kafka-connect-offsets",
                     },
-                    "nameOverride": CONNECTOR_NAME,
+                    "nameOverride": CONNECTOR_FULL_NAME,
                 },
             ),
             mocker.call.helm.uninstall(
                 namespace="test-namespace",
-                release_name="test-connector-with-long-name-0123456789abcdef-clean",
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 dry_run=False,
             ),
         ]
@@ -235,11 +238,11 @@ class TestKafkaSourceConnector(TestKafkaConnector):
             ),
             mocker.call.helm.uninstall(
                 namespace="test-namespace",
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 dry_run=False,
             ),
             mocker.call.helm.upgrade_install(
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 namespace="test-namespace",
                 chart="bakdata-kafka-connect-resetter/kafka-connect-resetter",
                 dry_run=False,
@@ -252,15 +255,15 @@ class TestKafkaSourceConnector(TestKafkaConnector):
                     "connectorType": "source",
                     "config": {
                         "brokers": "broker:9092",
-                        "connector": CONNECTOR_NAME,
+                        "connector": CONNECTOR_FULL_NAME,
                         "offsetTopic": "kafka-connect-offsets",
                     },
-                    "nameOverride": CONNECTOR_NAME,
+                    "nameOverride": CONNECTOR_FULL_NAME,
                 },
             ),
             mocker.call.helm.uninstall(
                 namespace="test-namespace",
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 dry_run=False,
             ),
         ]
@@ -311,11 +314,11 @@ class TestKafkaSourceConnector(TestKafkaConnector):
             ),
             mocker.call.helm.uninstall(
                 namespace="test-namespace",
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 dry_run=False,
             ),
             mocker.call.helm.upgrade_install(
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 namespace="test-namespace",
                 chart="bakdata-kafka-connect-resetter/kafka-connect-resetter",
                 dry_run=False,
@@ -328,15 +331,15 @@ class TestKafkaSourceConnector(TestKafkaConnector):
                     "connectorType": "source",
                     "config": {
                         "brokers": "broker:9092",
-                        "connector": CONNECTOR_NAME,
+                        "connector": CONNECTOR_FULL_NAME,
                         "offsetTopic": "kafka-connect-offsets",
                     },
-                    "nameOverride": CONNECTOR_NAME,
+                    "nameOverride": CONNECTOR_FULL_NAME,
                 },
             ),
             mocker.call.helm.uninstall(
                 namespace="test-namespace",
-                release_name=CONNECTOR_CLEAN_NAME,
+                release_name=CONNECTOR_CLEAN_FULL_NAME,
                 dry_run=False,
             ),
         ]